Move all cli wrappers to a single file
